@@ -17,52 +11,57 @@ version = project.properties["version"]
         ?.takeUnless { it == "unspecified" } ?: "1.0.0-SNAPSHOT"
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_1_7
-    targetCompatibility = JavaVersion.VERSION_1_7
-}
-
-repositories {
-    jcenter()
-    maven {
-        url = uri("https://dl.bintray.com/kotlin/kotlinx")
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(11))
     }
 }
 
+repositories {
+    mavenCentral()
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+
 dependencies {
-    val jacksonVersion = "2.9.7"
+    api(kotlin("stdlib"))
+    api("org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.8.0")
 
-    compile(kotlin("stdlib"))
-    compile("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.0.1")
 
-    testCompile("junit:junit:4.12")
-    testCompile("de.erichseifert.vectorgraphics2d:VectorGraphics2D:0.13")
+    testImplementation(kotlin("test"))
+    testImplementation("org.junit.jupiter:junit-jupiter-api:5.8.1")
+    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.8.1")
+    testImplementation("de.erichseifert.vectorgraphics2d:VectorGraphics2D:0.13")
 }
 
-val main by sourceSets
-
-val sourcesJar by tasks.creating(Jar::class) {
-    group = "build"
-    classifier = "sources"
-    from(main.allSource)
+dependencyLocking {
+    lockAllConfigurations()
 }
+
+val kotlinSourcesJar by tasks
 
 val dokkaJar by tasks.creating(Jar::class) {
     group = JavaBasePlugin.DOCUMENTATION_GROUP
     description = "Assembles Kotlin docs with Dokka"
-    classifier = "javadoc"
-    from(tasks["dokka"])
+    archiveClassifier = "javadoc"
+    from(tasks["dokkaJavadoc"])
 }
 
-tasks.named<Jar>("jar") {
+val mainJar by tasks.named<Jar>("jar")
+
+tasks.jar {
     dependsOn("generatePomFileForMavenJavaPublication")
     into("META-INF/maven/${project.group}/${project.name}") {
-        from(File(buildDir, "publications/mavenJava"))
+        from(layout.buildDirectory.dir("publications/mavenJava")) {
+            include("pom-default.xml")
+        }
         rename(".*", "pom.xml")
     }
 }
 
 jmh {
-    jmhVersion = "1.21"
+    jmhVersion = "1.37"
     jvmArgs = listOf("-Djmh.separateClasspathJAR=true")
 }
 
