@@ -2,14 +2,14 @@ package org.cikit.syslog
 
 sealed class ProgressiveCoderResult {
 
-    object Underflow : ProgressiveCoderResult()
+    data object Underflow : ProgressiveCoderResult()
 
-    object Overflow : ProgressiveCoderResult()
+    data object Overflow : ProgressiveCoderResult()
 
     class Malformed(val length: Int) : ProgressiveCoderResult()
 
     class Unmappable(val length: Int) : ProgressiveCoderResult()
 
-    object EOF : ProgressiveCoderResult()
+    data object EOF : ProgressiveCoderResult()
 
 }
